Moved stories files into folders of each component
