+++ frontend/src/components/ui/Button/Button.stories.js
@@ -1,4 +1,4 @@
-import Button from "../components/ui/Button/Button";
+import Button from "./Button";
 
 export default {
   title: "components/Button",
