+++ frontend/src/components/ui/Input/Input.stories.js
@@ -1,4 +1,4 @@
-import Input from "../components/ui/Input/Input";
+import Input from "./Input";
 
 export default {
   title: "components/Input",
